선박 등록 수정 중

+++ src/main/java/kroryi/dagon/repository/product/ProductImageRepository.java
@@ -3,6 +3,6 @@ package kroryi.dagon.repository.product;
 import kroryi.dagon.entity.product.ProductImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductImageRepository extends JpaRepository<ProductImage, Integer> {
+public interface ProductImageRepository extends JpaRepository<ProductImage, Long> {
 
 }
